#2 Базовые настройки среды разработки, разделение кода и подключение модулей
