+++ public/index.html
@@ -7,8 +7,8 @@
       content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
     />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="stylesheet" href="css/normalize.css" />
-    <link rel="stylesheet" href="css/style.css" />
+    <link rel="stylesheet" href="./css/normalize.css" />
+    <link rel="stylesheet" href="./css/style.css" />
     <title>TASKMANAGER</title>
   </head>
 
@@ -45,7 +45,7 @@
       <!-- Контент -->
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
 
   </body>
 </html>
