+++ src/components/load-button.js
@@ -1,0 +1,7 @@
+const createLoadButtonTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {createLoadButtonTemplate};
